hash join
The classic hash join algorithm for an inner join of two relations has the following steps:
Hash phase: Create a hash table for one of the two relations by applying a hash function to the join attribute of each row. Ideally we should create a hash table for the smaller relation, thus optimizing for creation time and memory size of the hash table.
Join phase: Scan the larger relation and find the relevant rows by looking in the hash table created before.